pyproject.toml

[tool.poetry]
name = "nlp"
version = "0.1.0"
description = ""
authors = ["Ramseths"]

[tool.poetry.dependencies]
python = "^3.9"
hydra-core = "^1.1.1"
pandas = "^2.2.2"
spacy = "^3.7.5"
nltk = "^3.8.1"
scikit-learn = "^1.5.1"
matplotlib = "^3.9.1"
seaborn = "^0.13.2"
gensim = "^4.3.2"
scipy = "1.12"
pyldavis = "^3.4.1"
mlflow = "^2.14.3"
transformers = "^4.42.4"
sentence-transformers = "^3.0.1"
datasets = "^2.20.0"
pre-commit = "2.18.0"
isort = "^5.13.2"

[tool.poetry.dev-dependencies]
pdoc3 = "^0.10.0"
pytest = "^6.2.5"
pre-commit = "^2.17.0"
ipykernel = "^6.28.0"
pandas-stubs = "^2.2.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

# Same as Black.
line-length = 88

[tool.ruff.lint]
ignore = ["E501"]
select = ["B","C","E","F","W","B9", "I", "Q"]

[tool.ruff.lint.mccabe]
max-complexity = 10

[tool.mypy]
ignore_missing_imports = true